+++ tracker/npm_package/CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [0.2.3] - 2025-06-17
-
-- Rename `Form Submission` system event to `Form: Submission`.
+- Add `logging` option
+- Improve `callback` option
+- Improve `fileDownloads` typing, export `DEFAULT_FILE_TYPES`
 
 ## [0.2.2] - 2025-06-16